requirements.txt
tensorflow-cpu
pandas 
numpy
ast
matplotlib
plotly